+++ heron/common/src/cpp/basics/mempool.cpp
@@ -17,7 +17,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 // Just defines the static member of Mempool class
 ////////////////////////////////////////////////////////////////////////////////
-#include "network/mempool.h"
+#include "basics/mempool.h"
 #include "basics/basics.h"
 
 MemPool<google::protobuf::Message>* __global_protobuf_pool__ =
