+++ heron/common/src/cpp/network/client.h
@@ -46,7 +46,6 @@
 #include "network/networkoptions.h"
 #include "network/network_error.h"
 #include "network/packet.h"
-#include "network/mempool.h"
 
 /*
  * Client class definition
